+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.13.0
+## 0.13.0 "Laconic Letters"
 
 + Added module `argparse` with class `LetterSubparsers()`, which allows pacman-style args by preprocessing them
   before feeding them to parse_args